GCF file split: firebase
